Move http service enabling before the reboot
Make sure http service is enable as part of the base system before
rebooting the node.

Leave node reboot in place for now but this shouln't be necessary
after a provision.  It will only work if the ansible playbook
is run external to the host.  The ansible playbook execution simply
ends when this is run local to the node, as is the case with vagrant
provisioning.