Add a dummy htpasswd file for fully working basic authn
Add a htpasswd file with user/pass for the vagrant user to
the repo.  Yes this is a password in git but it's just a
dummy account for user vagrant password vagrant.

It should be replaced by a more appropriate htpasswd that has
actual test cluster acounts in the future.

Also basic auth is just for dev anyway and should never be active
in prod.

There is likely a better way to do this, so future improvements
are welcome.