Add warewulf sync harness to ood node
Using warewulf synchronization is the easiest way to get
important files like the slurm.conf, munge.key, and hosts file
onto the ood node in order for it to participate in the cluster, ie.
schedule jobs.

The warewulf harness is added during node provisioning but since
the ood node is not provisioned by warewulf we need to add this
harness manually.  The /warewulf directory and associated cron.d
entry are the key to enabling this synchronization.

Once in place the ood node simply needs to be registered in the ohpc
warewulf db by ethernet address and the registered files will be
copied.

This feature is based on the discussion thread:

https://groups.io/g/OpenHPC-users/topic/27331998#2402

and in particular the process outlined here:

https://groups.io/g/OpenHPC-users/message/2084