Change the SELinux policy to targeted to fix the OOD reboot issue.
OOD would hang if you restart it after provisioning, this would be fixed by changing policy to targeted. The policy was set to permissive previously which wasn't one of the three available options for SELINUXTYPE in /etc/selinux/config. This fix would change the SELINUXTYPE in /etc/selinux/config to targeted.