Remove "System " prefix from default network manager profile names
Default network profile naming in CentOS-7 NetworkManager uses the
name  'System ethN' as the profile name for the 'ethN' interface.
The ansible framework uses a single variable for the profile name
interface name.   We change the default profile names to match
the interface name to match this convesion and make nmcli commands
more consistent with tranditional naming.