Change the playbook target to the headnode group name
Using the group name for headnode instead of the ohpc host name
from the inventory in the playbooks makes it easier to build
images with custom host names.
It's also slightly more correct to use the group name instead of
the host name and aligns with other uses of group name in playbooks.