Update build image with runtime instance name for slurm controller and db
The build-time setting of the slurm controller and db hostname cause
slurm startup to fail during deploy (ops) phase.  The ohpc_install_ops
role updates the slurm.conf from the template to include the current
names. The conditional avoids multiple writers to the file, since there
is one call to the code for each client node in the cluster due to the
current null resource config.