Deploy rabbitmq service and agents with enable_user_reg
Change to enable_user_reg flag for calling the rabbitmq server and agent
install playbooks so they are installed when the user_reg app
is installed.  These are dependancies of user_reg and should
be installed whenever that app is active.  The don't need to
have their own dedicated playbook flags.