Starts at Adding Device and IP Details for Virtual Box
A beginning attempt at adding details for the device names and IPs for Virtual Box VMs. 
In particular, it adds descriptions of those IPs in regards to their formatting.
There are also now notes of identified old documentation that might be able to be removed due to this new information.