Attempts Consolidation of Information Regarding Networks
The goal was to consolidate all information regarding details of the IPs and their network interfaces. This is a full consolidation with all information crammed into the list. 
This is all information relating to the VM content due to it's discussion of defaults.
More or less, this is a much more radical version of the "Addition_Mention_Defaults" branch.